Torch device (#9)
* Add model_device property to EmotionRecognizer class

It will allow to easily get the exact device the model's parameters are on (it is assumed all model parameters will be on a single device, it's not NASA space project)

* Make test_model method return accuracy

It now returns accuracy, rather than the model itself

* Add device conversion in make_one_prediction() method

Now the method always casts the data to the self.model_device device.
It is NOT done in forward() though, so that simply calling model(data) will not change data's device.

* Change main.py and train.py so that they always use CUDA when available.

Also during training, the DataLoaders pin memory if device other than cpu is used.